Fix timing, Much faster bit-banging

+++ VRC-Haptics-Firmware-ESP/src/globals.h
@@ -7,7 +7,7 @@
 namespace Haptics {
     // Volatile, non-static, user-denied variables
     struct Globals {
-        uint16_t ledcMotorVals[MAX_LEDC_MOTORS];
+        uint8_t ledcMotorVals[MAX_LEDC_MOTORS];
         uint16_t pcaMotorVals[MAX_I2C_MOTORS];
         uint16_t allMotorVals[MAX_LEDC_MOTORS + MAX_I2C_MOTORS];
         bool reinitLEDC;
